+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 * Update automatic release component information with N/A when pipelines have not run yet ([#1257](https://github.com/opendevstack/ods-jenkins-shared-library/pull/1257))
 * Enhance Helm strategy and support of Statefullset/Cronjob ([#1253](https://github.com/opendevstack/ods-jenkins-shared-library/pull/1253))
 * Add component information in automatic release close notes ([#1254](https://github.com/opendevstack/ods-jenkins-shared-library/pull/1254))
+* Improve Trivy Stage by being able to use severity thresholds ([#1262](https://github.com/opendevstack/ods-jenkins-shared-library/pull/1262))
 
 ### Fixed
 * Fail pipeline when deploying with issues in progress ([#1258](https://github.com/opendevstack/ods-jenkins-shared-library/pull/1258))